backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
supabase>=2.11.0
pyjwt[crypto]>=2.9.0
httpx>=0.28.0
nba_api>=1.5.0
pandas>=2.2.0
numpy>=2.1.0
scikit-learn>=1.6.0
pydantic-settings>=2.7.0
python-dotenv>=1.0.0
